refactor(TestConfig): convert to Singleton
